lib/init: do not look for vDSO support on Valgrind
There is no vDSO support with Valgrind. Detect when running on top of
such virtual machine using the Valgrind API, picking the fallback
clock_gettime() implementation if so.

This is not a problem since the EVL core does intercept and honor this
syscall directly from the out-of-band stage as needed. The incurred
overhead is negligible compared to what Valgrind already adds anyway.

Also, provide a placeholder for valgrind/valgrind.h which should be
picked when Valgrind is not installed on the build system.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>