include/valgrind: add original BSD-style licensed header
We may have no valgrind support in the SDK, but access to the virtual
machine in the runtime environment. So we should rather use the actual
RUN_ON_VALGRIND implementation from the original Valgrind header
(i.e. <valgrind/valgrind.h>), which is compatible license-wise.

Suppress the former placeholder in the same move.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>